requirements.txt
asgiref==3.5.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
Django==4.0.3
django-rest-framework==0.1.0
djangorestframework==3.13.1
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
lxml==4.8.0
mccabe==0.6.1
Pillow==9.1.0
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pytz==2022.1
requests==2.27.1
soupsieve==2.3.2
sqlparse==0.4.2
urllib3==1.26.9
